5300: Move old_system helpers into file, and rename them from labse
